Compile Plugin for Java 8
Java 8 is still supported to run Gradle. You may run Gradle with
Java 8 and still use another Java 9+ JDK to build Modules. So there
is no need to restrict this plugin to require Java 9+.

See #29 for the error you get when trying to run a build with this
plugin on Java 8.