Support 'group:name' notation and automatic version detection
The plugin is now able to identify a Jar by 'group:name' coordinates.
This information is used in the _merge Jar_ functionality to
automatically configure the 'javaModulesMergeJars' configuration
with all required dependencies.
The plugin can now also determine the version of the library to use
that version in the 'module-info.class'.

This all makes the DSL in 'extraJavaModuleInfo {}' significantly easier
to use.